# bytes: use bytealg.MakeNoZero in Buffer.growSlice